Move BackTracker to lexer
The BackTracker object is very specific to the lexer, so this commit moves
the interface and mock to the lexer.
