+++ common/character.go
@@ -58,82 +58,3 @@ func (m *MockCharStream) Next() (Char, error) {
 
 	return Char{}, args.Error(1)
 }
-
-// TrackAll is a special value for the max argument to
-// BackTracker.SetMax that indicates the desire to track all
-// characters.
-const TrackAll = -1
-
-// BackTracker is an interface for a backtracker, a CharStream that
-// also provides the ability to back up to an earlier character in the
-// stream.
-type BackTracker interface {
-	CharStream
-
-	// SetMax allows updating the maximum number of characters to
-	// allow backtracking over.  Setting a TrackAll value will
-	// allow all newly returned characters to be backtracked over.
-	// If the new value for max is less than the previous value,
-	// characters at the front of the backtracking queue will be
-	// discarded to bring the size down to max.
-	SetMax(max int)
-
-	// Accept accepts characters from the backtracking queue,
-	// leaving only the specified number of characters on the
-	// queue.
-	Accept(leave int)
-
-	// Len returns the number of characters saved so far on the
-	// backtracking queue.
-	Len() int
-
-	// Pos returns the position of the most recently returned
-	// character within the saved character list.
-	Pos() int
-
-	// BackTrack resets to the beginning of the backtracking
-	// queue.
-	BackTrack()
-}
-
-// MockBackTracker is a mock implementation of the BackTracker
-// interface.
-type MockBackTracker struct {
-	MockCharStream
-}
-
-// SetMax allows updating the maximum number of characters to allow
-// backtracking over.  Setting a TrackAll value will allow all newly
-// returned characters to be backtracked over.  If the new value for
-// max is less than the previous value, characters at the front of the
-// backtracking queue will be discarded to bring the size down to max.
-func (m *MockBackTracker) SetMax(max int) {
-	m.MethodCalled("SetMax", max)
-}
-
-// Accept accepts characters from the backtracking queue, leaving only
-// the specified number of characters on the queue.
-func (m *MockBackTracker) Accept(leave int) {
-	m.MethodCalled("Accept", leave)
-}
-
-// Len returns the number of characters saved so far on the
-// backtracking queue.
-func (m *MockBackTracker) Len() int {
-	args := m.MethodCalled("Len")
-
-	return args.Int(0)
-}
-
-// Pos returns the position of the most recently returned character
-// within the saved character list.
-func (m *MockBackTracker) Pos() int {
-	args := m.MethodCalled("Pos")
-
-	return args.Int(0)
-}
-
-// BackTrack resets to the beginning of the backtracking queue.
-func (m *MockBackTracker) BackTrack() {
-	m.MethodCalled("BackTrack")
-}
