+++ common/character_test.go
@@ -43,54 +43,3 @@ func TestMockCharStreamNextNotNil(t *testing.T) {
 	assert.Same(t, assert.AnError, err)
 	assert.Equal(t, Char{Rune: 'c'}, result)
 }
-
-func TestMockBackTrackerImplementBackTracker(t *testing.T) {
-	assert.Implements(t, (*BackTracker)(nil), &MockBackTracker{})
-}
-
-func TestMockBackTrackerSetMax(t *testing.T) {
-	obj := &MockBackTracker{}
-	obj.On("SetMax", 42)
-
-	obj.SetMax(42)
-
-	obj.AssertExpectations(t)
-}
-
-func TestMockBackTrackerAccept(t *testing.T) {
-	obj := &MockBackTracker{}
-	obj.On("Accept", 42)
-
-	obj.Accept(42)
-
-	obj.AssertExpectations(t)
-}
-
-func TestMockBackTrackerLen(t *testing.T) {
-	obj := &MockBackTracker{}
-	obj.On("Len").Return(42)
-
-	result := obj.Len()
-
-	assert.Equal(t, 42, result)
-	obj.AssertExpectations(t)
-}
-
-func TestMockBackTrackerPos(t *testing.T) {
-	obj := &MockBackTracker{}
-	obj.On("Pos").Return(42)
-
-	result := obj.Pos()
-
-	assert.Equal(t, 42, result)
-	obj.AssertExpectations(t)
-}
-
-func TestMockBackTrackerBackTrack(t *testing.T) {
-	obj := &MockBackTracker{}
-	obj.On("BackTrack")
-
-	obj.BackTrack()
-
-	obj.AssertExpectations(t)
-}
