+++ lexer/classify.go
@@ -16,8 +16,6 @@ package lexer
 
 import (
 	"github.com/stretchr/testify/mock"
-
-	"github.com/hydralang/ptk/common"
 )
 
 // Classifier represents a character classification tool.  A
@@ -29,11 +27,11 @@ type Classifier interface {
 	// wrapped in a BackTracker and determines one or more
 	// recognizers to extract a token or a set of tokens from the
 	// lexer input.
-	Classify(state State, str common.BackTracker) []Recognizer
+	Classify(state State, str BackTracker) []Recognizer
 
 	// Error is called by the lexer state if all recognizers
 	// returned by Classify return without success.
-	Error(state State, str common.BackTracker)
+	Error(state State, str BackTracker)
 }
 
 // MockClassifier is a mock implementation of the Classifier
@@ -45,7 +43,7 @@ type MockClassifier struct {
 // Classify takes a lexer state and the character stream wrapped in a
 // BackTracker and determines one or more recognizers to extract a
 // token or a set of tokens from the lexer input.
-func (m *MockClassifier) Classify(state State, str common.BackTracker) []Recognizer {
+func (m *MockClassifier) Classify(state State, str BackTracker) []Recognizer {
 	args := m.MethodCalled("Classify", state, str)
 
 	if tmp := args.Get(0); tmp != nil {
@@ -57,6 +55,6 @@ func (m *MockClassifier) Classify(state State, str common.BackTracker) []Recogni
 
 // Error is called by the lexer state if all recognizers returned by
 // Classify return without success.
-func (m *MockClassifier) Error(state State, str common.BackTracker) {
+func (m *MockClassifier) Error(state State, str BackTracker) {
 	m.MethodCalled("Error", state, str)
 }
