+++ lexer/recognizer_test.go
@@ -18,8 +18,6 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-
-	"github.com/hydralang/ptk/common"
 )
 
 func TestMockRecognizerImplementsRecognizer(t *testing.T) {
@@ -28,7 +26,7 @@ func TestMockRecognizerImplementsRecognizer(t *testing.T) {
 
 func TestMockRecognizerRecognize(t *testing.T) {
 	s := &MockState{}
-	str := &common.MockBackTracker{}
+	str := &MockBackTracker{}
 	obj := &MockRecognizer{}
 	obj.On("Recognize", s, str).Return(true)
